# docs/api/openapi.yaml
{
  "openapi": "3.1.1",
  "info": {
    "title": "StudentOS API",
    "description": "Personal academic operating system — REST API for web and mobile clients.\n\nAuthentication uses HS256 JWT access tokens (short-lived) and refresh tokens stored in the database. Attach the access token as `Authorization: Bearer <token>` or via the `x-access-token` header.\n\nAll authenticated endpoints return the standard envelope from the shared contract: `{ success: true, data: T, meta?: ... }` on success and `{ success: false, error: { code, message } }` on failure.\n\n**Ownership model:** Every resource (academic year, semester, course, task, note, study session, goal, etc.) belongs to a single user. Users can only see and modify their own resources. Unauthorized access returns `FORBIDDEN`. Missing resources return `NOT_FOUND`.",
    "version": "0.1.0",
    "contact": {
      "name": "StudentOS",
      "url": "https://github.com/elyasdemamu/StudentOS"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Base path for all endpoints"
    }
  ],
...
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new account",
        "description": "Creates a new student account. The password is hashed with Argon2id (never stored in plain text). Returns the public user profile (no password hash).",
        "operationId": "register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse_UserPublic" }
              }
            }
          },
          "409": {
            "description": "Email already in use",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Log in",
        "description": "Authenticates with email + password and returns an access token (JWT), its lifetime in seconds, and the public user profile. For web clients, also sets an HTTP-only `refresh_token` cookie.",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "headers": {
              "Set-Cookie": {
                "description": "HTTP-only refresh_token cookie (web clients only)",
                "schema": {
                  "type": "string",
                  "example": "refresh_token=eyJ...; Path=/api/v1/auth; HttpOnly; Secure; SameSite=Lax"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "401": {
            "description": "Invalid email or password",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Refresh access token",
        "description": "Uses a valid refresh token to issue a new access token + refresh token pair (rotation). Invalidates the old refresh token. Web clients also get a new HTTP-only cookie.",
        "operationId": "refresh",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token pair refreshed",
            "headers": {
              "Set-Cookie": {
                "description": "New HTTP-only refresh_token cookie",
                "schema": {
                  "type": "string",
                  "example": "refresh_token=eyJ...; Path=/api/v1/auth; HttpOnly; Secure; SameSite=Lax"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RefreshResponse" }
              }
            }
          },
          "401": {
            "description": "Invalid or expired refresh token",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Log out",
        "description": "Invalidates the refresh token so it cannot be reused. Clears the HTTP-only refresh cookie for web clients.",
        "operationId": "logout",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged out",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse_LoggedOut" }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get current user profile",
        "description": "Returns the public profile of the authenticated user.",
        "operationId": "getMe",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse_UserPublic" }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "HS256 JWT access token. Send as `Authorization: Bearer <token>` or `x-access-token: <token>`."
      }
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": ["email", "password", "firstName", "lastName"],
        "properties": {
          "email": { "type": "string", "format": "email", "example": "elyas@studentos.dev" },
          "password": { "type": "string", "minLength": 8, "format": "password", "example": "correct-horse-battery-staple" },
          "firstName": { "type": "string", "minLength": 1, "maxLength": 100, "example": "Elyas" },
          "lastName": { "type": "string", "minLength": 1, "maxLength": 100, "example": "Demamu" },
          "university": { "type": "string", "maxLength": 200, "nullable": true, "example": "Adis Abeba Science and Technology" },
          "department": { "type": "string", "maxLength": 200, "nullable": true, "example": "Computer Science" },
          "academicYear": { "type": "string", "nullable": true, "example": "2026-2027" },
          "timezone": { "type": "string", "default": "UTC" }
        },
        "additionalProperties": false
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "format": "password" }
        },
        "additionalProperties": false
      },
      "RefreshRequest": {
        "type": "object",
        "required": ["refreshToken"],
        "properties": {
          "refreshToken": { "type": "string", "description": "The refresh token from a previous login or refresh response." }
        },
        "additionalProperties": false
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string", "description": "Short-lived JWT access token." },
          "expiresIn": { "type": "integer", "description": "Access token lifetime in seconds." },
          "user": { "$ref": "#/components/schemas/UserPublic" }
        }
      },
      "RefreshResponse": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string" },
          "expiresIn": { "type": "integer" },
          "user": { "$ref": "#/components/schemas/UserPublic" }
        }
      },
...
      "UserPublic": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "university": { "type": "string", "nullable": true },
          "department": { "type": "string", "nullable": true },
          "academicYear": { "type": "string", "nullable": true },
          "timezone": { "type": "string" },
          "profilePicture": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" }
        },
        "additionalProperties": false,
        "description": "Public user representation. Never includes password hash."
      },
...
    }
  }
}